+++ src/pages/map/index.tsx
@@ -3,32 +3,10 @@ import Head from "next/head"
 import { Box } from "@mantine/core"
 import axios from "axios"
 
+import { loadCss } from "@/lib/loadCss"
+import { loadScript } from "@/lib/loadScript"
+
 import type { NextPage } from "next"
-
-const loadScript = (url: string, onload: () => void) => {
-  // check if it is already loaded
-  if (!Array.from(document.scripts).find((s) => s.src == url)) {
-    const head = document.getElementsByTagName("head")[0] as HTMLElement
-    const script = document.createElement("script")
-    script.setAttribute("type", "text/javascript")
-    script.setAttribute("src", url)
-    script.addEventListener("load", onload)
-
-    head.appendChild(script)
-  }
-}
-
-const loadCss = (url: string) => {
-  // check if it is already loaded
-  if (!Array.from(document.getElementsByTagName("link")).find((l) => l.href == url)) {
-    const head = document.getElementsByTagName("head")[0] as HTMLElement
-    const link = document.createElement("link")
-    link.setAttribute("rel", "stylesheet")
-    link.setAttribute("href", url)
-
-    head.appendChild(link)
-  }
-}
 
 const mapSetUp = () => {
   // show map on Box
